@@ -10,33 +10,38 @@
   "license": "MIT",
   "dependencies": {
     "body-parser": "1.18.2",
-    "express": "4.16.2",
-    "request": "2.83.0",
-    "react": "15.1.0",
-    "react-dom": "15.1.0",
-    "socket.io": "2.0.4",
-    "socket.io-client": "2.0.4",
     "d3": "3.5.17",
-    "react-faux-dom": "2.5.0"
+    "express": "4.16.2",
+    "prop-types": "15.6.0",
+    "react": "16.2.0",
+    "react-dom": "16.2.0",
+    "react-faux-dom": "2.5.0",
+    "request": "2.83.0",
+    "socket.io": "2.0.4",
+    "socket.io-client": "2.0.4"
   },
   "scripts": {
     "test": "npm run test-client && npm run test-server",
-    "test-client": "mocha test/client/**/*.js --compilers js:babel-core/register",
+    "test-client": "mocha --compilers js:babel-core/register test/client/**/*.js",
     "test-server": "mocha test/server/test-main.js",
     "start": "nodemon dist/main.js --port=3000",
     "build": "gulp",
     "build-stage": "gulp stage",
+    "lint": "eslint --ext .js --ext .jsx src/** && eslint test/**",
     "coverage": "npm run c-server && npm run c-combine",
     "c-server": "./node_modules/istanbul/lib/cli.js cover --dir ./coverage/server ./node_modules/mocha/bin/_mocha test/server/test-main.js -- -R spec",
     "c-combine": "./node_modules/istanbul/lib/cli.js report --dir ./coverage --include **/*coverage.json json lcov"
   },
   "devDependencies": {
-    "babel-core": "6.9.1",
-    "babel-loader": "6.2.4",
+    "@cygns/eslint-config": "2.0.0",
+    "babel-core": "6.26.0",
     "babel-preset-es2015": "6.9.0",
-    "babel-preset-react": "6.5.0",
-    "babelify": "7.3.0",
-    "browserify": "13.0.1",
+    "babel-preset-react": "6.24.1",
+    "babelify": "8.0.0",
+    "browserify": "16.1.0",
+    "eslint": "4.18.1",
+    "eslint-plugin-import": "2.9.0",
+    "eslint-plugin-react": "7.7.0",
     "expect": "1.20.1",
     "expect-jsx": "2.6.0",
     "gulp": "3.9.1",
@@ -44,15 +49,15 @@
     "gulp-gzip": "1.4.0",
     "gulp-sass": "2.3.2",
     "gulp-sourcemaps": "1.6.0",
-    "gulp-uglify": "1.5.3",
+    "gulp-uglify": "3.0.0",
     "gulp-util": "3.0.7",
     "istanbul": "0.4.3",
     "mocha": "3.0.1",
-    "nodemon": "1.9.2",
-    "react-addons-test-utils": "15.1.0",
-    "vinyl-buffer": "1.0.0",
-    "vinyl-source-stream": "1.1.0",
-    "watchify": "3.7.0"
+    "nodemon": "1.15.1",
+    "react-test-renderer": "16.2.0",
+    "vinyl-buffer": "1.0.1",
+    "vinyl-source-stream": "2.0.0",
+    "watchify": "3.10.0"
   },
   "engines": {
     "node": "8.9.3"
