@@ -1,37 +1,41 @@
 /**
- * Copyright (c) Joe McIntyre, 2016
+ * Copyright (c) Joe McIntyre, 2016-2018
  * license: MIT (https://github.com/fcc-joemcintyre/stocktracker/LICENSE.txt)
  */
-import React from 'react';
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 
 /**
  * Outstanding requests and error status display.
  */
-export default class Range extends React.Component {
+export class Range extends Component {
   constructor (props) {
     super (props);
     this.state = {
-      months: props.months
+      months: props.months,
     };
     this.onRangeSelected = this.onRangeSelected.bind (this);
   }
 
   onRangeSelected (months) {
     if (months !== this.state.months) {
-      this.setState ({months: months});
+      this.setState ({ months });
       this.props.onRangeChanged (months);
     }
   }
 
-  render() {
-    let spans = [];
-    for (let value of [1, 3, 6, 12, 36]) {
+  render () {
+    const spans = [];
+    for (const value of [1, 3, 6, 12, 36]) {
       spans.push (
-        <span key={value}
-          className = {(value === this.state.months) ? 'selected' : ''}
-          onClick={this.onRangeSelected.bind (this, value)}>
-          {(value < 12) ? value + 'm' : (value / 12) + 'y'}
-        </span>);
+        <span
+          key={value}
+          className={(value === this.state.months) ? 'selected' : ''}
+          onClick={() => { this.onRangeSelected (value); }}
+        >
+          {(value < 12) ? `${value}m` : `${(value / 12)}y`}
+        </span>
+      );
     }
 
     return (
@@ -44,11 +48,11 @@ export default class Range extends React.Component {
 
 Range.propTypes = {
   // months in range
-  months: React.PropTypes.number,
+  months: PropTypes.number,
   // function to call on range change
-  onRangeChanged: React.PropTypes.func.isRequired
-}
+  onRangeChanged: PropTypes.func.isRequired,
+};
 
 Range.defaultProps = {
-  months: 12
-}
+  months: 12,
+};
