@@ -1,16 +1,15 @@
 /**
- * Copyright (c) Joe McIntyre, 2016
+ * Copyright (c) Joe McIntyre, 2016-2018
  * license: MIT (https://github.com/fcc-joemcintyre/stocktracker/LICENSE.txt)
  */
-'use strict';
 
 /**
  * Calculate a date in the past by a number of months.
- * @param date Starting date
- * @param monthsAgo Number of months to go back
- * @return Date provided minus number of months
+ * @param {Date} date Starting date
+ * @param {number} monthsAgo Number of months to go back
+ * @return {Date} Date provided minus number of months
  */
-function dateOffset (date, monthsAgo) {
+export function dateOffset (date, monthsAgo) {
   if (monthsAgo === 0) {
     return date;
   }
@@ -18,7 +17,7 @@ function dateOffset (date, monthsAgo) {
   let year = date.getFullYear () - Math.floor (monthsAgo / 12);
   let month = date.getMonth () - (monthsAgo % 12);
   if (month < 0) {
-    year --;
+    year -= 1;
     month += 12;
   }
 
@@ -26,11 +25,9 @@ function dateOffset (date, monthsAgo) {
   let day = date.getDate () + 1;
   let result = new Date (year, month, day);
   while (result.getMonth () !== month) {
-    day --;
+    day -= 1;
     result = new Date (year, month, day);
   }
 
   return (result);
 }
-
-exports.dateOffset = dateOffset;
