@@ -1,8 +1,7 @@
 /**
- * Copyright (c) Joe McIntyre, 2016
+ * Copyright (c) Joe McIntyre, 2016-2018
  * license: MIT (https://github.com/fcc-joemcintyre/stocktracker/LICENSE.txt)
  */
-'use strict';
 const request = require ('request');
 
 // socket.io and stock data
@@ -12,7 +11,8 @@ let stocks = [];
 /**
  * Initialize. Setup Socket.io connection listener, sending current current
  * data to newly connecting client.
- * @param _io Socket.io instance
+ * @param {Object} _io Socket.io instance
+ * @return {void}
  */
 function init (_io) {
   io = _io;
@@ -24,6 +24,9 @@ function init (_io) {
 
 /**
  * Register interest in a stock to track.
+ * @param {Object} req Request
+ * @param {Object} res Response
+ * @return {void}
  */
 function registerStock (req, res) {
   let result;
@@ -32,7 +35,7 @@ function registerStock (req, res) {
     // add if symbol not already active
     symbol = symbol.toString ().trim ().toUpperCase ();
     let found = false;
-    for (let stock of stocks) {
+    for (const stock of stocks) {
       if (symbol === stock.symbol) {
         found = true;
         break;
@@ -41,15 +44,18 @@ function registerStock (req, res) {
     if (! found) {
       addStock (symbol);
     }
-    result = {errorCode:0};
+    result = { errorCode: 0 };
   } else {
-    result = {errorCode:1, message:'Symbol missing'};
+    result = { errorCode: 1, message: 'Symbol missing' };
   }
   res.status (200).json (result);
 }
 
 /**
  * Deregister interest in a stock to track.
+ * @param {Object} req Request
+ * @param {Object} res Response
+ * @return {void}
  */
 function deregisterStock (req, res) {
   let result;
@@ -65,63 +71,64 @@ function deregisterStock (req, res) {
         break;
       }
     }
-    result = {errorCode:0};
+    result = { errorCode: 0 };
   } else {
-    result = {errorCode:1, message:'Symbol missing'};
+    result = { errorCode: 1, message: 'Symbol missing' };
   }
   res.status (200).json (result);
 }
 
 /**
  * Add a stock, fetching its historical data (async)
- * @param symbol Stock trading symbol
+ * @param {string} symbol Stock trading symbol
+ * @return {void}
  */
 function addStock (symbol) {
   process.nextTick (() => {
     // calculate start / end dates (3 year history)
-    let date = new Date ();
+    const date = new Date ();
     date.setFullYear (date.getFullYear () - 3);
-    let startDate = `${date.getFullYear ()}-${date.getMonth () + 1}-${date.getDate ()}`;
-    let dates= `start_date=${startDate}`;
+    const startDate = `${date.getFullYear ()}-${date.getMonth () + 1}-${date.getDate ()}`;
+    const dates = `start_date=${startDate}`;
 
     // construct url and retrieve data
-    let key = process.env.QKEY;
-    let keyParam = (key) ? `&api_key=${key}` : '';
-    let base = `https://www.quandl.com/api/v3/datasets/WIKI/${symbol}.json`;
-    let url = `${base}?order=asc&${dates}${keyParam}`;
+    const key = process.env.QKEY;
+    const keyParam = (key) ? `&api_key=${key}` : '';
+    const base = `https://www.quandl.com/api/v3/datasets/WIKI/${symbol}.json`;
+    const url = `${base}?order=asc&${dates}${keyParam}`;
     request.get (url, (err, res, body) => {
       if (! err) {
         if (res.statusCode === 200) {
-          let data = JSON.parse (body);
-          let index = data.dataset.name.indexOf ('(');
-          let name = (index === -1) ? data.dataset.name : data.dataset.name.substring (0, index - 1);
+          const data = JSON.parse (body);
+          const index = data.dataset.name.indexOf ('(');
+          const name = (index === -1) ? data.dataset.name : data.dataset.name.substring (0, index - 1);
           // add stock to set of tracked stocks, broadcast to connected clients
           stocks.push ({
             status: 0,
-            symbol: symbol,
-            name: name,
-            data: data.dataset.data
+            symbol,
+            name,
+            data: data.dataset.data,
           });
           broadcast ();
           console.log (`${symbol} now being tracked`);
         } else {
           stocks.push ({
             status: res.statusCode,
-            symbol: symbol,
-            name: undefined,
-            data: undefined
+            symbol,
+            name: null,
+            data: null,
           });
           console.log ('Fetch error: status code ', res.statusCode);
-          let data = JSON.parse (body);
+          const data = JSON.parse (body);
           console.log ('  Detailed msg', data);
           broadcast ();
         }
       } else {
         stocks.push ({
           status: 1,
-          symbol: symbol,
-          name: undefined,
-          data: undefined
+          symbol,
+          name: null,
+          data: null,
         });
         console.log ('Fetch error:', err);
         broadcast ();
@@ -132,6 +139,7 @@ function addStock (symbol) {
 
 /**
  * Broadcast tracked stocks to all connected clients (async)
+ * @return {void}
  */
 function broadcast () {
   process.nextTick (() => {
@@ -140,7 +148,7 @@ function broadcast () {
       io.emit ('update', JSON.stringify (stocks));
 
       // clean up any error records after each broadcast
-      stocks = stocks.filter (a => { return a.status === 0; });
+      stocks = stocks.filter (a => (a.status === 0));
     }
   });
 }
