@@ -1,6 +1,4 @@
-'use strict';
-const server = require ('../../dist/server');
-
+/* eslint-disable global-require */
 describe ('test-main', function () {
   describe ('test-cmd', function () {
     require ('./test_cmd');
