@@ -2,82 +2,81 @@
  * Copyright (c) Joe McIntyre, 2016
  * license: MIT (https://github.com/fcc-joemcintyre/stocktracker/LICENSE.txt)
  */
-'use strict';
 const assert = require ('assert');
 const processCommand = require ('../../dist/cmd').processCommand;
 
 describe ('Test command line processing', function () {
   describe ('empty command', function () {
     it ('should not fail', function () {
-      let cmd = processCommand ([]);
-      assert.deepStrictEqual (cmd, {code:0, exit:false, port:3000 });
+      const cmd = processCommand ([]);
+      assert.deepStrictEqual (cmd, { code: 0, exit: false, port: 3000 });
     });
   });
 
   describe ('invalid standalone option', function () {
     it ('should fail with code 1', function () {
-      let cmd = processCommand (['-j']);
-      assert.deepStrictEqual (cmd, {code:1, exit:true, port:3000 });
+      const cmd = processCommand (['-j']);
+      assert.deepStrictEqual (cmd, { code: 1, exit: true, port: 3000 });
     });
   });
 
   describe ('invalid settings option', function () {
     it ('should fail with code 1', function () {
-      let cmd = processCommand (['-j=foo.js']);
-      assert.deepStrictEqual (cmd, {code:1, exit:true, port:3000 });
+      const cmd = processCommand (['-j=foo.js']);
+      assert.deepStrictEqual (cmd, { code: 1, exit: true, port: 3000 });
     });
   });
 
   describe ('proper port argument', function () {
     it ('should succeed', function () {
-      let cmd = processCommand (['-p=2000']);
-      assert.deepStrictEqual (cmd, {code:0, exit:false, port:2000 });
+      const cmd = processCommand (['-p=2000']);
+      assert.deepStrictEqual (cmd, { code: 0, exit: false, port: 2000 });
     });
   });
 
   describe ('port out of range (negative)', function () {
     it ('should fail', function () {
-      let cmd = processCommand (['-p=-1']);
-      assert.deepStrictEqual (cmd, {code:1, exit:true, port:'-1' });
+      const cmd = processCommand (['-p=-1']);
+      assert.deepStrictEqual (cmd, { code: 1, exit: true, port: '-1' });
     });
   });
 
   describe ('port out of range (positive)', function () {
     it ('should fail', function () {
-      let cmd = processCommand (['-p=200000']);
-      assert.deepStrictEqual (cmd, {code:1, exit:true, port:'200000' });
+      const cmd = processCommand (['-p=200000']);
+      assert.deepStrictEqual (cmd, { code: 1, exit: true, port: '200000' });
     });
   });
 
   describe ('port not an integer', function () {
     it ('should fail', function () {
-      let cmd = processCommand (['-p=2000.5']);
-      assert.deepStrictEqual (cmd, {code:1, exit:true, port:'2000.5' });
+      const cmd = processCommand (['-p=2000.5']);
+      assert.deepStrictEqual (cmd, { code: 1, exit: true, port: '2000.5' });
     });
   });
 
   describe ('port not a number', function () {
     it ('should fail', function () {
-      let cmd = processCommand (['-p=ABC']);
-      assert.deepStrictEqual (cmd, {code:1, exit:true, port:'ABC' });
+      const cmd = processCommand (['-p=ABC']);
+      assert.deepStrictEqual (cmd, { code: 1, exit: true, port: 'ABC' });
     });
   });
 
   describe ('unary help command', function () {
     it ('should succeed', function () {
       let cmd = processCommand (['-h']);
-      assert.deepStrictEqual (cmd, {code:0, exit:true, port:3000 });
+      assert.deepStrictEqual (cmd, { code: 0, exit: true, port: 3000 });
       cmd = processCommand (['--help']);
-      assert.deepStrictEqual (cmd, {code:0, exit:true, port:3000 });
+      assert.deepStrictEqual (cmd, { code: 0, exit: true, port: 3000 });
     });
   });
 
   describe ('help in command', function () {
     it ('should succeed', function () {
       let cmd = processCommand (['-p=2000', '-h']);
-      assert.deepStrictEqual (cmd, {code:0, exit:true, port:2000 });
+      assert.deepStrictEqual (cmd, { code: 0, exit: true, port: 2000 });
       cmd = processCommand (['-p=2000', '--help']);
-      assert.deepStrictEqual (cmd, {code:0, exit:true, port:2000 });
+      assert.deepStrictEqual (cmd, { code: 0, exit: true, port: 2000 });
     });
   });
 });
